test-clone/python-backend/requirements.txt
google-genai>=0.3.0
websockets>=12.0
pyaudio>=0.2.11
Pillow>=9.0.0
asyncio